Moved ignitus-Testimonial to HomePage.
