+++ src/ignitus-HomePage/ignitus-Testimonial/sagas.js
@@ -1,7 +1,7 @@
 
 import { effects } from 'redux-saga';
 import * as t from './actionTypes';
-import * as api from '../ignitus-Api';
+import * as api from '../../ignitus-Api';
 
 const {
   call, put, takeLatest, all,
