+++ src/ignitus-Routes/publicRoutes.js
@@ -6,7 +6,7 @@ import Footer from "../ignitus-Footer";
 import Contributors from "../ignitus-HomePage/ignitus-Contributors";
 import GetStarted from "../ignitus-HomePage/ignitus-GetStarted";
 import Introduction from "../ignitus-HomePage/ignitus-Introduction";
-import Testimonial from "../ignitus-Testimonial";
+import Testimonial from "../ignitus-HomePage/ignitus-Testimonial";
 import WhatWeProvide from "../ignitus-WhatWeDo";
 import Partner from "../ignitus-HomePage/ignitus-Partners";
