+++ src/rootSagas.ts
@@ -1,6 +1,6 @@
 
 import { all } from 'redux-saga/effects';
-import Testimonial from './ignitus-Testimonial';
+import Testimonial from './ignitus-HomePage/ignitus-Testimonial';
 import studentSignUp from './ignitus-Authentication/ignitus-StudentSignUp';
 import studentLogin from './ignitus-Authentication/ignitus-StudentLogin';
 import contributors from './ignitus-Team';
